Instead of parsing the input while also checking the sequence, I created a method that converts the string input to an integer array for easier use
Part 2 was quite difficult. I first attempted to count the number of problems found and then allow 1 problem before determining the sequence to be unsafe, but it was not the correct answer. I wanted to prioritize efficiency, but in the end I knew no better choice than to constantly exclude values until a safe sequence was found.