Day 3 Part 2 solution
Part 2 actually wasn't as difficult as I anticipated. I initially predicted it would be something like allowing spaces in the expressions or whatnot. This solution works, however I do recognize some flaws in this solution. For example, if a 'do()' or 'don't()' command appears after the last 'mul()' command on a line, it will not be included as the program stops recording commands after the final 'mul()' command on a line. Luckily my particular input did not have any instances like that, but I don't know what other people's input is looking like.